project3/part2/requirements.txt
boto3
facenet-pytorch
imutils
numpy
opencv-python-headless
Pillow
torch
torchvision